@@ -1,19 +1,17 @@
 name: barras
-description: A simple and customizable barcode scanning package for Android and iOS. It uses AVCaptureSession in iOS and ZXing in Android.
-version: 0.1.0-nullsafety.1
+description: A simple and customizable barcode scanning package for Android and iOS. It uses CameraX on Android and AVFoundation on iOS.
+version: 1.0.0
 homepage: https://github.com/deadblit/barras
 
 environment:
-  sdk: ">=2.13.4 <3.0.0"
-  flutter: ">=2.2.3"
+  sdk: ">=2.17.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
-  qrcode_forked: ^1.0.1
-  torch_controller: ^0.4.1
   flutter_beep: ^1.0.0
+  mobile_scanner: ^3.4.1
 
 dev_dependencies:
   flutter_test:
