@@ -2,8 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_beep/flutter_beep.dart';
-import 'package:qrcode_forked/qrcode_forked.dart';
-import 'package:torch_controller/torch_controller.dart';
+import 'package:mobile_scanner/mobile_scanner.dart';
 
 import 'barcode_reader_overlay_painter.dart';
 
@@ -39,10 +38,12 @@ class BarcodeReaderPage extends StatefulWidget {
 }
 
 class _BarcodeReaderPageState extends State<BarcodeReaderPage> {
-  QRCaptureController _captureController = QRCaptureController();
+  MobileScannerController _captureController = MobileScannerController(
+    detectionSpeed: DetectionSpeed.normal,
+    facing: CameraFacing.back,
+    torchEnabled: false,
+  );
 
-  bool _hasTorch = false;
-  bool _isTorchOn = false;
   bool _isBorderVisible = false;
   Timer? _borderFlashTimer;
 
@@ -64,23 +65,11 @@ class _BarcodeReaderPageState extends State<BarcodeReaderPage> {
         });
       }
     }
-    TorchController().hasTorch.then((value) {
-      setState(() {
-        _hasTorch = value ?? false;
-      });
-    });
-
-    _captureController.onCapture((data) {
-      _captureController.pause();
-      if (widget.successBeep) {
-        FlutterBeep.beep();
-      }
-      Navigator.of(context).pop(data);
-    });
   }
 
   @override
   void dispose() {
+    _captureController.dispose();
     _borderFlashTimer?.cancel();
     super.dispose();
   }
@@ -90,7 +79,7 @@ class _BarcodeReaderPageState extends State<BarcodeReaderPage> {
     return Scaffold(
       body: Stack(
         alignment: Alignment.center,
-        children: <Widget>[
+        children: [
           _buildCaptureView(),
           _buildViewfinder(context),
           _buildButtonBar(),
@@ -100,7 +89,12 @@ class _BarcodeReaderPageState extends State<BarcodeReaderPage> {
   }
 
   Widget _buildCaptureView() {
-    return QRCaptureView(
+    return MobileScanner(
+      onDetect: (barcodes) async {
+        await _captureController.stop();
+        if (widget.successBeep) FlutterBeep.beep();
+        Navigator.of(context).pop(barcodes.barcodes.first.rawValue);
+      },
       controller: _captureController,
     );
   }
@@ -128,7 +122,7 @@ class _BarcodeReaderPageState extends State<BarcodeReaderPage> {
         child: Row(
           mainAxisSize: MainAxisSize.max,
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: <Widget>[
+          children: [
             _buildTorchButton(),
             _buildCancelButton(),
           ],
@@ -138,37 +132,34 @@ class _BarcodeReaderPageState extends State<BarcodeReaderPage> {
   }
 
   Widget _buildTorchButton() {
-    return (_hasTorch)
+    return (_captureController.hasTorch)
         ? IconButton(
-            icon: Icon(
-              (_isTorchOn) ? Icons.flash_on : Icons.flash_off,
-              color: widget.buttonColor,
+            icon: ValueListenableBuilder(
+              builder: (context, state, child) {
+                switch (state as TorchState) {
+                  case TorchState.on:
+                    return Icon(Icons.flash_on, color: widget.buttonColor);
+                  case TorchState.off:
+                    return Icon(Icons.flash_off, color: widget.buttonColor);
+                }
+              },
+              valueListenable: _captureController.torchState,
             ),
             onPressed: () {
-              if (_isTorchOn) {
-                _captureController.torchMode = CaptureTorchMode.off;
-              } else {
-                _captureController.torchMode = CaptureTorchMode.on;
-              }
-
-              setState(() {
-                _isTorchOn = !_isTorchOn;
-              });
+              _captureController.toggleTorch();
             },
           )
-        : Container(
-            width: 10,
-            height: 10,
-          );
+        : Container(width: 10, height: 10);
   }
 
   Widget _buildCancelButton() {
-    return FlatButton(
+    return TextButton(
       onPressed: () {
-        _captureController.pause();
         Navigator.of(context).pop();
       },
-      textColor: widget.buttonColor,
+      style: TextButton.styleFrom(
+        foregroundColor: widget.buttonColor,
+      ),
       child: Text(widget.cancelButtonText),
     );
   }
